Moved r.js to its own folder
